Add user nick/account/hostname tracking
This uses a local fs db to persist data. It introduces core.async for
more robust async communication and scheduling. Tracking is done
passively, both when bayaz enters a channel and when a new user joins a
channel bayaz is in. Tracking is severely throttled to avoid network
spam.

Currently, none of this info is surfaced via commands.